Fix create recipe record registring (#146)
* Fix typo

* Move records controller
